requirements.txt
Pillow
torchvision
openai-clip
torch
ptvsd
dotenv
debugpy
google-api-python-client
openai
numpy
bokeh
matplotlib
pre-commit
tokenize-rt
astunparse
